backend/requirements.txt
fastapi
uvicorn[standard]
pythainlp
python-crfsuite
scikit-learn
pydantic>=2
python-docx
email-validator
python-decouple
motor
passlib[bcrypt]
bcrypt==3.2.2
PyJWT
numpy
python-multipart
PyPDF2
pymupdf
docx2txt
networkx
google-genai